* tag 'v0.8.0': (625 commits)
  Fix release instruction for making docs
  Update docversion for 0.8
  Update version number for 0.8
  Add release notes for version 0.8
  DOC: Fix typo in remove_small_objects docstring.
  Improve performance of normalization
  Change color of circles for better visibility
  Remove unnecessary statement breaks
  Improve performance and usage of data types of hough circle transformation
  STY: Minor cleanup.
  Add specific error messages and exceptions
  Removing unnecessary typedefs for uint16_t in MSVC build
  Change data type of index variables
  Fix missed rename in example
  Add raise and labeled image tests
  Rename function to remove_small_objects
  Improve docstring for remove_small_connected_components
  Modify example to use remove_small_connected_components
  Add tests for remove_small_connected_components
  Add examples to docstring
  ...

+++ skimage/_shared/transform.pxd
@@ -2,4 +2,4 @@ cimport numpy as cnp
 
 
 cdef float integrate(cnp.ndarray[float, ndim=2,  mode="c"] sat,
-                     int r0, int c0, int r1, int c1)
+                     Py_ssize_t r0, Py_ssize_t c0, Py_ssize_t r1, Py_ssize_t c1)
